fix tests

+++ ct-app/test/components/test_messages.py
@@ -5,7 +5,7 @@ def test_create_message():
     message = MessageFormat("random_relayer")
     assert message.relayer == "random_relayer"
     assert message.timestamp is not None
-    assert isinstance(message.timestamp, float)
+    assert isinstance(message.timestamp, int)
 
 def test_parse_message():
     encoded = MessageFormat("random_relayer")
